Use signatures to identify Chainguard images.
We can't diff non-Chainguard images, so when images are hosted outside
of `cgr.dev` we need some way of figuring out what is or isn't a
Chainguard image.

Previously I was using the vendor annotation but this will be inherited
by images that are based on Chainguard images. Those images aren't
diffable because they (most likely) won't have the expected SBOM.

This changes the workflow to use signatures to identify Chainguard
images.

Signed-off-by: Rob Best <rob.best@chainguard.dev>